pyproject.toml
[project]
name = "feishu-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.39.0",
]
